// src/openshift/gen/model/comGithubOpenshiftApiProjectV1ProjectSpec.ts
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.22.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* ProjectSpec describes the attributes on a Project
*/
export class ComGithubOpenshiftApiProjectV1ProjectSpec {
    /**
    * Finalizers is an opaque list of values that must be empty to permanently remove object from storage
    */
    'finalizers'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "finalizers",
            "baseName": "finalizers",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return ComGithubOpenshiftApiProjectV1ProjectSpec.attributeTypeMap;
    }
}