// src/openshift/gen/model/comGithubOpenshiftApiProjectV1ProjectStatus.ts
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.22.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { V1NamespaceCondition } from './v1NamespaceCondition';

/**
* ProjectStatus is information about the current status of a Project
*/
export class ComGithubOpenshiftApiProjectV1ProjectStatus {
    /**
    * Represents the latest available observations of the project current state.
    */
    'conditions'?: Array<V1NamespaceCondition>;
    /**
    * Phase is the current lifecycle phase of the project
    */
    'phase'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "conditions",
            "baseName": "conditions",
            "type": "Array<V1NamespaceCondition>"
        },
        {
            "name": "phase",
            "baseName": "phase",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ComGithubOpenshiftApiProjectV1ProjectStatus.attributeTypeMap;
    }
}